+++ proofs/clock.php
@@ -12,9 +12,9 @@ use Innmind\BlackBox\Set;
 return static function() {
     yield proof(
         'All models are always within bounds',
-        given(Set\Either::any(
+        given(Set::either(
             PointInTime::any(),
-            Set\Call::of(static fn() => Clock::live()->now()),
+            Set::call(static fn() => Clock::live()->now()),
         )),
         static function($assert, $point) {
             $assert
@@ -96,8 +96,8 @@ return static function() {
     yield proof(
         'Clock::at() returns nothing for invalid strings',
         given(
-            Set\Unicode::strings(),
-            Set\Elements::of(
+            Set::strings()->unicode(),
+            Set::of(
                 Format::cookie(),
                 Format::iso8601(),
                 Format::rfc1036(),
@@ -121,7 +121,7 @@ return static function() {
         'Clock::ofFormat()->at()',
         given(
             PointInTime::any(),
-            Set\Elements::of(
+            Set::of(
                 Format::iso8601(),
                 new class implements Format\Custom {
                     public function normalize(): Format
@@ -155,7 +155,11 @@ return static function() {
 
     yield proof(
         'Each call to Clock::now() is ahead of the previous',
-        given(Set\Nullable::of(Set\Integers::between(1, 2_000_000))), // up to 2 seconds
+        given(
+            Set::integers()
+                ->between(1, 2_000_000) // up to 2 seconds
+                ->nullable(),
+        ),
         static function($assert, $microsecond) {
             $clock = Clock::live();
             $start = $clock->now();